Move LyricEnhancer to Singelton
